+++ aux_codec.js
@@ -13,8 +13,8 @@ or implied.
 *
 * Repository: gve_devnet_webex_devices_executive_room_multi_aux_switching_macro
 * Macro file: aux_codec
-* Version: 1.0.14
-* Released: November 14, 2023
+* Version: 1.0.15
+* Released: November 19, 2023
 * Latest RoomOS version tested: 11.10.1.8
 *
 * Macro Author:      	Gerardo Chaves